requirements.txt
alembic==0.8.10
amqp==2.1.4
appdirs==1.4.0
billiard==3.5.0.2
blinker==1.4
celery==4.0.2
click==6.7
dominate==2.3.1
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.9.5
kombu==4.0.2
Mako==1.0.6
MarkupSafe==0.23
olefile==0.44
packaging==16.8
passlib==1.7.1
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python-editor==1.0.3
pytz==2016.10
raven==5.32.0
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.5
vine==1.1.3
visitor==0.1.3
Werkzeug==0.11.15
WTForms==2.1
